refactor(forum): use role-based access policies

+++ public/request/forum-topic/delete.php
@@ -1,11 +1,11 @@
 <?php
 
-use App\Enums\Permissions;
 use App\Models\ForumTopic;
+use App\Models\User;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Validator;
 
-if (!authenticateFromCookie($username, $permissions, $userDetails, Permissions::Registered)) {
+if (!authenticateFromCookie($username, $permissions, $userDetails)) {
     return back()->withErrors(__('legacy.error.permissions'));
 }
 
@@ -15,6 +15,13 @@ $input = Validator::validate(Arr::wrap(request()->post()), [
 
 /** @var ForumTopic $topic */
 $topic = ForumTopic::find($input['topic']);
+
+$userModel = User::firstWhere('User', $username);
+
+if (!$userModel->can('delete', $topic)) {
+    return back()->withErrors(__('legacy.error.permissions'));
+}
+
 $topic->delete();
 
 return back()->with('success', __('legacy.success.delete'));
